2023.07.13 - 2 (showBytes() was fixed)

+++ app/src/main/kotlin/ir/mahdiparastesh/instatools/view/UiTools.kt
@@ -228,13 +228,13 @@ object UiTools {
     fun Context.showBytes(length: Long): String {
         val units = resources.getStringArray(R.array.bytes)
         var unit = 0
-        var nominalSize = length
-        while ((nominalSize / 1024L) > 1) {
-            nominalSize /= 1024L
+        var nominalSize = length.toDouble()
+        while ((nominalSize / 1024.0) > 1.0) {
+            nominalSize /= 1024.0
             unit++
             if (unit == units.size - 1) break
         }
-        return units[unit].format(nominalSize)
+        return units[unit].format(nominalSize.toInt())
     }
 
     /** @return a datetime text to be used in a file name. */
